+++ app/javascript/engine/actions/UndeployAction.ts
@@ -1,4 +1,4 @@
-import { Coordinate, Direction } from "../../utilities/commonTypes";
+import { Coordinate } from "../../utilities/commonTypes";
 import { formatCoordinate, formatNation } from "../../utilities/graphics";
 import DeployState from "../control/state/DeployState";
 import Game from "../Game";
@@ -6,9 +6,8 @@ import { GameActionPath, GameActionData, GameActionReinforcementUnit } from "../
 import Unit from "../Unit";
 import BaseAction from "./BaseAction";
 
-export default class DeployAction extends BaseAction {
+export default class UndeployAction extends BaseAction {
   target: Coordinate;
-  orientation: Direction;
   rTurn: number;
   rKey: string;
   rId: string;
@@ -17,7 +16,6 @@ export default class DeployAction extends BaseAction {
     super(data, game, index)
 
     this.validate(data.data.deploy)
-    // Validate will already error out if data is missing, but the linter can't tell
     const deploy = data.data.deploy as GameActionReinforcementUnit[]
     this.validate(deploy[0])
     this.validate(data.data.path)
@@ -25,25 +23,36 @@ export default class DeployAction extends BaseAction {
     this.validate(path[0])
 
     this.target = new Coordinate(path[0].x, path[0].y)
-    this.orientation = path[0].facing ?? 1
     this.rKey = deploy[0].key
     this.rTurn = deploy[0].turn
     this.rId = deploy[0].id
   }
 
-  get type(): string { return "deploy" }
+  get type(): string { return "undeploy" }
 
   get htmlValue(): string {
     const name = this.player === 1 ?
       this.game.scenario.alliedReinforcements[this.rTurn][this.rKey].counter.name :
       this.game.scenario.axisReinforcements[this.rTurn][this.rKey].counter.name
-    return `deployed ${formatNation(this.game, this.player)} ` +
-      `unit: ${formatNation(this.game, this.player,name)} to ${formatCoordinate(this.target)}`
+    return `undeployed ${formatNation(this.game, this.player)} ` +
+      `unit: ${formatNation(this.game, this.player,name)} from ${formatCoordinate(this.target)}`
   }
 
   get undoPossible() { return true }
 
   mutateGame(): void {
+    const scenario = this.game.scenario
+
+    if (this.player === 1) {
+      scenario.replaceAlliedReinforcement(this.rTurn, this.rKey)
+    } else {
+      scenario.replaceAxisReinforcement(this.rTurn, this.rKey)
+    }
+    this.map.removeCounter(this.target, this.rId)
+    this.game.setGameState(new DeployState(this.game, this.rTurn, this.rKey))
+  }
+
+  undo(): void {
     const scenario = this.game.scenario
     const map = scenario.map
 
@@ -54,26 +63,7 @@ export default class DeployAction extends BaseAction {
     if (!uf.isFeature) {
       (uf as Unit).playerNation = this.player === 1 ? scenario.alliedFactions[0] : scenario.axisFactions[0]
     }
-    if (uf.rotates) {
-      uf.facing = this.orientation
-      if (!uf.isFeature) {
-        const unit = uf as Unit
-        unit.turretFacing = this.orientation
-      }
-    }
     uf.id = this.rId
     map.addCounter(this.target, uf)
-  }
-
-  undo(): void {
-    const scenario = this.game.scenario
-
-    if (this.player === 1) {
-      scenario.replaceAlliedReinforcement(this.rTurn, this.rKey)
-    } else {
-      scenario.replaceAxisReinforcement(this.rTurn, this.rKey)
-    }
-    this.map.removeCounter(this.target, this.rId)
-    this.game.setGameState(new DeployState(this.game, this.rTurn, this.rKey))
   }
 }